@@ -1,7 +1,16 @@
 import { Component, EventEmitter, Input, Output, ViewEncapsulation } from '@angular/core';
 
-export enum DoubleButtonActive { RightButton, LeftButton }
+/**
+ * Identifies the active button of a DoubleButtonComponent.
+ */
+export enum DoubleButtonActive {
+  RightButton = 'RightButton',
+  LeftButton = 'LeftButton',
+}
 
+/**
+ * Bar with 2 buttons, one active and other inactive. Used to select between two options.
+ */
 @Component({
   selector: 'app-double-button',
   templateUrl: './double-button.component.html',
@@ -9,15 +18,20 @@ export enum DoubleButtonActive { RightButton, LeftButton }
   encapsulation: ViewEncapsulation.Emulated,
 })
 export class DoubleButtonComponent {
-  @Input() rightButtonText: any;
-  @Input() leftButtonText: any;
+  @Input() rightButtonText: string;
+  @Input() leftButtonText: string;
   @Input() activeButton: DoubleButtonActive;
+  // Allows to add classes to the component.
   @Input() className = '';
+  // If true, when the user clicks one of the buttons the newly selected button will not
+  // be selected automatically. Instead, the control will just send the event indicating the
+  // click and the "activeButton" property will have to be changed for the clicked button to
+  // be selected.
   @Input() changeActiveButtonManually = false;
-  @Output() onStateChange = new EventEmitter();
-  ButtonState = DoubleButtonActive;
+  @Output() onStateChange = new EventEmitter<DoubleButtonActive>();
+  ButtonStates = DoubleButtonActive;
 
-  onRightClick() {
+  onRightButtonClicked() {
     if (this.activeButton === DoubleButtonActive.LeftButton) {
       if (!this.changeActiveButtonManually) {
         this.activeButton = DoubleButtonActive.RightButton;
@@ -26,7 +40,7 @@ export class DoubleButtonComponent {
     }
   }
 
-  onLeftClick() {
+  onLeftButtonClicked() {
     if (this.activeButton === DoubleButtonActive.RightButton) {
       if (!this.changeActiveButtonManually) {
         this.activeButton = DoubleButtonActive.LeftButton;
