@@ -1,17 +1,22 @@
 import { Component, Inject, OnDestroy, ViewChild, ElementRef } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA, MatDialog } from '@angular/material/dialog';
+import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+
 import { HwWalletService } from '../../../../services/hw-wallet.service';
 import { ChildHwDialogParams } from '../hw-options-dialog/hw-options-dialog.component';
 import { HwDialogBaseComponent } from '../hw-dialog-base.component';
-import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { ChangeNameComponent, ChangeNameData } from '../../../pages/wallets/change-name/change-name.component';
 import { MsgBarService } from '../../../../services/msg-bar.service';
-import { OperationError, HWOperationResults } from '../../../../utils/operation-error';
+import { OperationError } from '../../../../utils/operation-error';
 import { processServiceError } from '../../../../utils/errors';
 import { WalletsAndAddressesService } from '../../../../services/wallet-operations/wallets-and-addresses.service';
 import { WalletBase } from '../../../../services/wallet-operations/wallet-objects';
 import { HardwareWalletService } from '../../../../services/wallet-operations/hardware-wallet.service';
 
+/**
+ * Modal window used to add a new device to the wallet list. This modal window was created
+ * for being oppenend by the hw wallet options modal window.
+ */
 @Component({
   selector: 'app-hw-added-dialog',
   templateUrl: './hw-added-dialog.component.html',
@@ -23,10 +28,11 @@ export class HwAddedDialogComponent extends HwDialogBaseComponent<HwAddedDialogC
   form: FormGroup;
   maxHwWalletLabelLength = HwWalletService.maxLabelLength;
 
+  // Saves the initial label of the device, to know if the user tried to change it.
   private initialLabel: string;
 
   constructor(
-    @Inject(MAT_DIALOG_DATA) public data: ChildHwDialogParams,
+    @Inject(MAT_DIALOG_DATA) private data: ChildHwDialogParams,
     public dialogRef: MatDialogRef<HwAddedDialogComponent>,
     hwWalletService: HwWalletService,
     private formBuilder: FormBuilder,
@@ -36,7 +42,10 @@ export class HwAddedDialogComponent extends HwDialogBaseComponent<HwAddedDialogC
     private hardwareWalletService: HardwareWalletService,
   ) {
     super(hwWalletService, dialogRef);
+
+    // Add the device to the wallets list.
     this.operationSubscription = this.walletsAndAddressesService.createHardwareWallet().subscribe(wallet => {
+      // Update the security warnings.
       this.operationSubscription = this.hardwareWalletService.getFeaturesAndUpdateData(wallet).subscribe(() => {
         this.wallet = wallet;
         this.initialLabel = wallet.label;
@@ -45,8 +54,9 @@ export class HwAddedDialogComponent extends HwDialogBaseComponent<HwAddedDialogC
           label: [wallet.label, Validators.required],
         });
 
-        this.closeIfHwDisconnected = false;
         this.currentState = this.states.Finished;
+
+        // Request the data and state of the hw wallet options modal window to be refreshed.
         this.data.requestOptionsComponentRefresh();
 
         setTimeout(() => this.input.nativeElement.focus());
@@ -56,16 +66,9 @@ export class HwAddedDialogComponent extends HwDialogBaseComponent<HwAddedDialogC
 
   private processError(err: OperationError) {
     err = processServiceError(err);
-    if (err.type === HWOperationResults.Disconnected) {
-      this.closeModal();
+    this.processHwOperationError(err);
 
-      return;
-    }
-
-    this.showResult({
-      text: err.translatableErrorMsg,
-      icon: this.msgIcons.Error,
-    });
+    // Make the hw wallet options modal window show the error msg.
     this.data.requestOptionsComponentRefresh(err.translatableErrorMsg);
   }
 
@@ -76,10 +79,12 @@ export class HwAddedDialogComponent extends HwDialogBaseComponent<HwAddedDialogC
 
   saveNameAndCloseModal() {
     if (this.form.value.label === this.initialLabel) {
+      // If no change was made to the label, just close the window.
       this.closeModal();
     } else {
       this.msgBarService.hide();
 
+      // Open the appropiate component to change the device label.
       const data = new ChangeNameData();
       data.wallet = this.wallet;
       data.newName = this.form.value.label;
