@@ -1,10 +1,16 @@
 import { Component, Inject } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+
 import { HwWalletService } from '../../../../services/hw-wallet.service';
 import { ChildHwDialogParams } from '../hw-options-dialog/hw-options-dialog.component';
 import { HwDialogBaseComponent } from '../hw-dialog-base.component';
-import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 
+/**
+ * Allows to load a seed on a seedless device or to let the user to enter a seed to check
+ * if it is equal to the one on the device (if a wallet object is included in the params). This
+ * modal window was created for being oppenend by the hw wallet options modal window.
+ */
 @Component({
   selector: 'app-hw-restore-seed-dialog',
   templateUrl: './hw-restore-seed-dialog.component.html',
@@ -12,6 +18,8 @@ import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 })
 export class HwRestoreSeedDialogComponent extends HwDialogBaseComponent<HwRestoreSeedDialogComponent> {
   form: FormGroup;
+  // If true, the seed entered by the user will not be loaded on the device, the operation will
+  // just compare it to the one on the device.
   justCheckingSeed: boolean;
 
   constructor(
@@ -35,6 +43,7 @@ export class HwRestoreSeedDialogComponent extends HwDialogBaseComponent<HwRestor
     this.operationSubscription = this.hwWalletService.recoverMnemonic(this.form.controls['words'].value, this.justCheckingSeed).subscribe(
       () => {
         if (!this.justCheckingSeed) {
+          // Request the data and state of the hw wallet options modal window to be refreshed.
           this.data.requestOptionsComponentRefresh();
           this.closeModal();
         } else {
@@ -44,7 +53,7 @@ export class HwRestoreSeedDialogComponent extends HwDialogBaseComponent<HwRestor
           });
         }
       },
-      err => this.processResult(err),
+      err => this.processHwOperationError(err),
     );
   }
 }
