@@ -1,10 +1,15 @@
 import { Component, Inject } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+
 import { HwWalletService } from '../../../../services/hw-wallet.service';
 import { ChildHwDialogParams } from '../hw-options-dialog/hw-options-dialog.component';
 import { HwDialogBaseComponent } from '../hw-dialog-base.component';
 import { WalletsAndAddressesService } from '../../../../services/wallet-operations/wallets-and-addresses.service';
 
+/**
+ * Allows to wipe the device and to remove the wallet from the list. This modal window was
+ * created for being oppenend by the hw wallet options modal window.
+ */
 @Component({
   selector: 'app-hw-wipe-dialog',
   templateUrl: './hw-wipe-dialog.component.html',
@@ -22,16 +27,21 @@ export class HwWipeDialogComponent extends HwDialogBaseComponent<HwWipeDialogCom
   ) {
     super(hwWalletService, dialogRef);
 
+    // If no wallet was send as part of the data, the option for removing the wallet from
+    // the wallet list is not shown and no wallet is removed, as there is no way to know which
+    // wallet to remove. This is for wiping unknown devices, mainly ones blocked with a PIN code.
     if (!data.wallet) {
       this.showDeleteFromList = false;
       this.deleteFromList = false;
     }
   }
 
+  // Changes the option which indicates if the wallet must be removed from the wallet list.
   setDeleteFromList(event) {
     this.deleteFromList = event.checked;
   }
 
+  // Wipes the device.
   requestWipe() {
     this.currentState = this.states.Processing;
 
@@ -41,12 +51,16 @@ export class HwWipeDialogComponent extends HwDialogBaseComponent<HwWipeDialogCom
           text: 'hardware-wallet.general.completed',
           icon: this.msgIcons.Success,
         });
+
+        // Request the data and state of the hw wallet options modal window to be refreshed.
         this.data.requestOptionsComponentRefresh();
+
+        // Remove the wallet from the list, if requested.
         if (this.deleteFromList) {
           this.walletsAndAddressesService.deleteHardwareWallet(this.data.wallet.id);
         }
       },
-      err => this.processResult(err),
+      err => this.processHwOperationError(err),
     );
   }
 }
