@@ -3,16 +3,23 @@ import { MatDialogRef, MatDialog, MatDialogConfig } from '@angular/material/dial
 
 import { LanguageData, LanguageService } from '../../../services/language.service';
 
+/**
+ * Allows to change the language displayed by the UI.
+ */
 @Component({
   selector: 'app-select-language',
   templateUrl: './select-language.component.html',
   styleUrls: ['./select-language.component.scss'],
 })
 export class SelectLanguageComponent implements OnInit {
-
   languages: LanguageData[];
   disableDismiss: boolean;
 
+  /**
+   * Opens the modal window. Please use this function instead of opening the window "by hand".
+   * @param disableClose Disables the options for closing the modal window without
+   * selecting a langhuage.
+   */
   public static openDialog(dialog: MatDialog, disableClose = false): MatDialogRef<SelectLanguageComponent, any> {
     const config = new MatDialogConfig();
     config.autoFocus = false;
@@ -33,6 +40,10 @@ export class SelectLanguageComponent implements OnInit {
   }
 
   closePopup(language: LanguageData = null) {
-    this.dialogRef.close(language ? language.code : undefined);
+    if (language) {
+      this.languageService.changeLanguage(language.code);
+    }
+
+    this.dialogRef.close();
   }
 }
